refactor(step4): 로또 가격, 구매 티켓 개수, 로또 구매 총액 로직 위치 변경
